% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract}
\alias{extract}
\title{Extract design from an OptimalALT object}
\usage{
extract(object)
}
\arguments{
\item{object}{An object of class `OptimalALT`.}
}
\value{
The extracted design.
}
\description{
Extract design from an OptimalALT object
}
