% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design.R
\name{extract_design}
\alias{extract_design}
\title{Extract Optimal Design}
\usage{
extract_design(optimal_alt)
}
\arguments{
\item{optimal_alt}{An object of class `OptimalALT`.}
}
\value{
A matrix containing stress levels and allocated proportion of 
 the design, each row corresponds to stress levels of each factor, 
 while the last row is their proportions.
}
\description{
Extract design from an OptimalALT object
}
\examples{

design_info <- set_design_info(
    k_levels=2, 
    j_factor=1, 
    n_unit=300, 
    censor_time=183, 
    p=0.1, 
    use_cond=0, 
    sigma=0.6
)

pso_info <- pso_setting(
    n_swarm=5, 
    max_iter=10, 
    early_stopping=5, 
    tol=0.0001
)

example1_locally <- find_optimal_alt(
    design_type="locally", 
    distribution="weibull",
    design_info=design_info, 
    pso_info=pso_info, 
    coef = c(0.001, 0.9),
    highest_level=TRUE,
    verbose=TRUE, 
    n_threads = 1)
    
opt_design <- extract_design(example1_locally)   

  
}
