% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.DesignInfo}
\alias{print.DesignInfo}
\title{Print a DesignInfo Object}
\usage{
\method{print}{DesignInfo}(x, ...)
}
\arguments{
\item{x}{An object of class \code{DesignInfo}, typically returned
by \code{\link{set_design_info}}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original \code{DesignInfo} object.
}
\description{
Prints the design specifications stored in a \code{DesignInfo} object.
}
\examples{
\dontrun{
design_info <- set_design_info(...)
print(design_info)
} 

}
