% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.OptimalALT}
\alias{print.OptimalALT}
\title{Print an OptimalALT Object}
\usage{
\method{print}{OptimalALT}(x, ...)
}
\arguments{
\item{x}{An object of class \code{OptimalALT}, typically returned 
by \code{\link{find_optimal_alt}}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original \code{OptimalALT} object.
}
\description{
Prints a summary of the optimization results and optimality check 
 for an \code{OptimalALT} object.
}
\examples{
\dontrun{ 
result <- find_optimal_alt(...) 
print(result) 
}

}
