% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.OptimalityCheck}
\alias{print.OptimalityCheck}
\title{Print an OptimalityCheck Object}
\usage{
\method{print}{OptimalityCheck}(x, show_candidates = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{OptimalityCheck}, typically 
returned by \code{\link{check_optimality}}.}

\item{show_candidates}{Logical. If \code{TRUE}, the candidate model 
set used in the optimality check is printed. Default is \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns the original \code{OptimalityCheck} object.
}
\description{
Prints the results of an optimality check for an accelerated life
 test (ALT) design.
}
\examples{
\dontrun{
optimality <- check_optimality(...)
print(optimality)
}

}
