% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{update_optimality_check}
\alias{update_optimality_check}
\title{Update optimality check for an OptimalALT object}
\usage{
update_optimality_check(object, check)
}
\arguments{
\item{object}{An object of class `OptimalALT`}

\item{check}{An object of class `OptimalityCheck`}
}
\value{
An `OptimalityALT` object with updated optimality check information
}
\description{
Update optimality check for an OptimalALT object
}
