% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_optimality.R
\name{check_optimality}
\alias{check_optimality}
\title{Check Optimality for Optimal Design}
\usage{
check_optimality(best_design, model_set, design_info, seed = 42)
}
\arguments{
\item{best_design}{A matrix containing stress levels and allocated 
proportion of the design, each row corresponds to stress levels of 
each factor, while the last row is their proportions.}

\item{model_set}{A matrix of models, including parameters and distribution, 
that maximize the optimality criteria with the given best particle's position, 
each columns corresponds to model coefficients, while the last column 
is lifetime distribution.}

\item{design_info}{A `DesignInfo` object created by 
function \code{\link{set_design_info}}.}

\item{seed}{Seed for reproducibility}
}
\value{
An OptimalityCheck object containing optimality check results
}
\description{
Evaluates whether a design is optimal by the equivalence theorem.
}
\examples{
design_info <- set_design_info(
    k_levels=2, 
    j_factor=1, 
    n_unit=300, 
    censor_time=183, 
    p=0.1, 
    use_cond=0, 
    sigma=0.6
)
                               
best_design <- rbind(
  c(0.682, 1), 
  c(0.706, 0.294)
)

model_set <- rbind(
  c(0.01, 0.9, "weibull"),
  c(0.01, 0.99, "lognormal")
)

equi <- check_optimality (
    best_design = best_design, 
    model_set = model_set, 
    design_info = design_info
)
                          
print(equi)                          

}
\references{
\enumerate{
  \item Müller, C. H., & Pázman, A. (1998). Applications of necessary and sufficient conditions for maximin efficient designs. Metrika, 48, 1–19.
  \item Huang, M.-N. L., & Lin, C.-S. (2006). Minimax and maximin efficient designs for estimating the location-shift parameter of parallel models with dual responses. Journal of Multivariate Analysis, 97(1), 198–210.
}
}
