% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design.R
\name{extract_design}
\alias{extract_design}
\title{Extract design from an OptimalALT object}
\usage{
extract_design(optimal_alt)
}
\arguments{
\item{optimal_alt}{An object of class `OptimalALT`.}
}
\value{
The extracted design.
}
\description{
Extract design from an OptimalALT object
}
