% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaxALT.R
\name{find_optimal_alt}
\alias{find_optimal_alt}
\title{Find Optimal ALT Design Using Hybrid Algorithm}
\usage{
find_optimal_alt(
  design_type,
  distribution,
  design_info,
  pso_info,
  coef = NULL,
  coef_lower = NULL,
  coef_upper = NULL,
  init_values = NULL,
  highest_level = FALSE,
  n_threads = 1,
  verbose = TRUE,
  seed = 42
)
}
\arguments{
\item{design_type}{Character. One of c("locally", "minimax").}

\item{distribution}{Character. Failure distribution, one of 
c("weibull, "lognormal").}

\item{design_info}{A `DesignInfo` object from \code{\link{set_design_info}} containing 
design specifications.}

\item{pso_info}{A `PSOInfo` object from \code{\link{pso_setting}} defining 
PSO hyperparameters.}

\item{coef}{Optional. Fixed model coefficients. 
Required if \code{design_type = "locally"}.}

\item{coef_lower}{Optional. Lower bounds for model parameters. 
Required if \code{design_type = "minimax"}.}

\item{coef_upper}{Optional. Upper bounds for model parameters. 
Required if \code{design_type = "minimax"}.}

\item{init_values}{Optional. An `InitialValue` object of initial values 
from \code{\link{initialize_values}}.}

\item{highest_level}{Logical. Whether the highest stress level of the 
generated design is the upper bound of stress range \code{x_h}. 
Default value is \code{FALSE}.}

\item{n_threads}{Integer. Number of threads for parallel processing.}

\item{verbose}{Logical. If \code{TRUE}, print optimization progress.}

\item{seed}{Integer. Seed for reproducibility}
}
\value{
An object of class OptimalALT
}
\description{
Runs hybrid algorithm combining PSO and Nelder-Mead to find the optimal design of accelerated life test (ALT).
}
\examples{

design_info <- set_design_info(
    k_levels=3, 
    j_factor=1, 
    n_unit=300, 
    censor_time=183, 
    p=0.1, 
    use_cond=0, 
    sigma=0.6
)

pso_info <- pso_setting(
    n_swarm=5, 
    max_iter=10, 
    early_stopping=5, 
    tol=0.00001
)

res <- find_optimal_alt(
    design_type="minimax", 
    distribution="lognormal", 
    design_info=design_info, 
    pso_info=pso_info, 
    coef_lower=c(10^-6, 0.7),
    coef_upper=c(10^-3, 0.99), 
    highest_level = TRUE,
    verbose = FALSE,
    n_threads = 1
)

print(res)
summary(res)
plot(res, x_l=0, x_h=1)

}
\references{
\enumerate{
  \item Chen P (2024). _globpso: Particle Swarm Optimization Algorithms and Differential Evolution for Minimization Problems_. R package version 1.2.1, <https://github.com/PingYangChen/globpso>.
  \item Kennedy, J., & Eberhart, R. (1995). Particle swarm optimization. In Proceedings of the IEEE International Conference on Neural Networks (ICNN) (Vol. 4, pp. 1942–1948).
  \item Lee, I. C., Chen, R. B., Wong, W. K., (in press). Optimal Robust Strategies for Accelerated Life Tests and Fatigue Testing of Polymer Composite Materials. Annals of Applied Statistics. <https://imstat.org/journals-and-publications/annals-of-applied-statistics/annals-of-applied-statistics-next-issues/>
  \item Meeker, W. Q., & Escobar, L. A. (1998). Statistical methods for reliability data. New York: Wiley-Interscience.
  \item Nelder, J. A. and Mead, R. (1965). A simplex algorithm for function minimization. Computer Journal, 7, 308--313. 10.1093/comjnl/7.4.308.
}
}
