% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaxALT-package.R
\docType{package}
\name{minimaxALT-package}
\alias{minimaxALT}
\alias{minimaxALT-package}
\title{minimaxALT: Generate Optimal Designs of Accelerated Life Test using PSO-based Algorithm}
\description{
A computationally efficient solution for generating optimal experimental designs in Accelerated Life Testing (ALT) with multiple factors. Leveraging a Particle Swarm Optimization (PSO)-based hybrid algorithm, the package identifies optimal test plans that minimize estimation variance under specified failure models and stress profiles.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hoanglinh171/minimaxALT}
  \item Report bugs at \url{https://github.com/hoanglinh171/minimaxALT/issues}
}

}
\author{
\strong{Maintainer}: Hoai-Linh Hoang \email{hoailinh.hoang17@gmail.com}

}
\keyword{internal}
